Force single core untilize on BH (#15272)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/14061

### Problem description
The multicore untilize op is currently broken on BH.

### What's changed
- Temporary change to force untilize to use the single core op on BH to
unblock bringup efforts.
- Change to the Python reference implementation of untilize so that it
doesn't modify the dtype when untilizing. I came across this problem
while debugging #14061 and decided to roll the fix into this PR.

### Checklist
- [ ] [Post commit CI
passes](https://github.com/tenstorrent/tt-metal/actions/runs/11941777312)
- [ ] [Model regression CI testing
passes](https://github.com/tenstorrent/tt-metal/actions/runs/11941778139)
- [ ] [Device performance regression CI testing
passes](https://github.com/tenstorrent/tt-metal/actions/runs/11941778941)
- [ ] [Blackhole post commit
passes](https://github.com/tenstorrent/tt-metal/actions/runs/11941930513)
- [ ] New/Existing tests provide coverage for changes

+++ models/utility_functions.py
@@ -508,9 +508,9 @@ def untilize(x):
     ) == 0, "The last two dimensions of the tensor must be divisible by 32"
 
     if isinstance(x, torch.Tensor):
-        ret = torch.zeros(x.shape)
+        ret = torch.zeros(x.shape, dtype=x.dtype)
     else:
-        ret = np.zeros(x.shape)
+        ret = np.zeros(x.shape, dtype=x.dtype)
 
     for B in range(x.shape[0]):
         for C in range(x.shape[1]):
